+++ src/SearchInput/SearchInput.tsx
@@ -5,12 +5,17 @@ import React, {
   ReactNode,
   forwardRef,
   useMemo,
+  useRef,
   useState,
 } from 'react'
+import styled from 'styled-components'
 import { Box } from '../Box'
+import { IconStrict } from '../IconStrict'
 import { Field } from '../fields/Field'
 import { CommonFieldProps } from '../fields/commonFieldTypes'
 import { Input, StyledFrontIcon } from '../fields/components/CommonInput'
+import { useOnClickOutside } from '../hooks'
+import { theme } from '../theme'
 import { useUniqueId } from '../utils/id'
 import { useControllableState } from '../utils/useControlledState'
 import { SearchOptions } from './components/SearchOptions'
@@ -37,9 +42,11 @@ export interface SearchInputProps extends CommonFieldProps {
   /**  optional callback to run when no results found */
   onNotFound?: (searchTerm: string) => void
   /**  optional Component to render when no results found */
-  notFoundComponent?: ReactNode
+  notFoundComponent?: (searchTerm: string) => ReactNode
   /**  optional boolean to show search icon */
   showIcon?: boolean
+  /**  optional boolean to show a clear search button */
+  clearSearch?: boolean
   /**  Optional callback to run on blur */
   onBlur?: (e: FocusEvent<HTMLInputElement>) => void
   /**  Optional default value for input */
@@ -57,9 +64,18 @@ export interface SearchInputProps extends CommonFieldProps {
 }
 
 const defaultFuzzySearchOptions = {
-  keys: ['label', 'value'],
+  keys: [
+    {
+      name: 'label',
+      weight: 0.6,
+    },
+    {
+      name: 'tags',
+      weight: 0.4,
+    },
+  ],
   findAllMatches: true,
-  minMatchCharLength: 2,
+  minMatchCharLength: 1,
   location: 0,
   threshold: 0.45,
   distance: 55,
@@ -85,10 +101,12 @@ export const SearchInput = forwardRef<HTMLInputElement, SearchInputProps>(
       resultsBorder = true,
       enableFuzzySearch = false,
       fuzzySearchOptions,
+      clearSearch,
       ...otherProps
     },
     ref,
   ) {
+    const wrapperRef = useRef(null)
     const id = useUniqueId(idProp)
     const [showOptions, setShowOptions] = useState(false)
     const [selectedValue, setSelectedValue] = useControllableState<
@@ -98,6 +116,28 @@ export const SearchInput = forwardRef<HTMLInputElement, SearchInputProps>(
       stateProp: value,
     })
     const [searchQuery, setSearchQuery] = useState<string | null>(null)
+    const [highlightedIndex, setHighlightedIndex] = useState(-1)
+
+    const selectedValueLabel = searchList.find(
+      (option) =>
+        option.label === selectedValue || option.value === selectedValue,
+    )?.label
+
+    const handleBlur = () => {
+      if (selectedValue) {
+        setSearchQuery(selectedValueLabel ?? null)
+      } else if (!selectedValue) {
+        setSearchQuery(null)
+      }
+    }
+
+    useOnClickOutside({
+      ref: wrapperRef,
+      callback: () => {
+        handleBlur()
+        setShowOptions(false)
+      },
+    })
 
     const fuse = useMemo(() => {
       return new Fuse(searchList, {
@@ -124,16 +164,9 @@ export const SearchInput = forwardRef<HTMLInputElement, SearchInputProps>(
       if (searchQuery !== null) {
         return searchQuery
       }
-
       if (selectedValue !== null) {
-        return (
-          searchList.find(
-            (option) =>
-              option.label === selectedValue || option.value === selectedValue,
-          )?.label || ''
-        )
+        return selectedValueLabel || ''
       }
-
       return ''
     }
 
@@ -142,11 +175,24 @@ export const SearchInput = forwardRef<HTMLInputElement, SearchInputProps>(
 
     const updateSearchQuery = (query: string | null) => {
       setSearchQuery(query)
+      setHighlightedIndex(-1)
 
       if (query === null) {
+        setSelectedValue(null)
         setShowOptions(false)
       } else {
-        setShowOptions(2 <= query.length)
+        setShowOptions(true)
+      }
+    }
+
+    const handleClick = () => {
+      setShowOptions(true)
+      if (searchQuery !== null) {
+        updateSearchQuery(searchQuery)
+        setShowOptions(true)
+      } else if (selectedValue !== null && searchQuery === null) {
+        setSearchQuery(selectedValueLabel || null)
+        setShowOptions(true)
       }
     }
 
@@ -157,52 +203,135 @@ export const SearchInput = forwardRef<HTMLInputElement, SearchInputProps>(
 
     const handleSelect = (nextValue: SearchInputItem): void => {
       updateSearchQuery(null)
-
       setSelectedValue(nextValue.label)
       onFound(nextValue.value)
     }
 
-    return (
-      <Field
-        className={className}
-        renderAsTitle={renderAsTitle}
-        htmlFor={id}
-        {...otherProps}
-      >
-        <Box flex alignItems="center" justifyContent="flex-start">
-          {showIcon && <StyledFrontIcon render="search" color="sesame" />}
-          <Input
-            id={id}
-            name={name}
-            ref={ref}
-            placeholder={placeholder}
-            $frontIcon={showIcon ? 'search' : undefined}
-            $fallbackStyle={fallbackStyle}
-            autoComplete="off"
-            value={displayedInputText}
-            onChange={handleInputChange}
-            selected={isSelected}
-            onBlur={(e) => {
-              if (displayedInputText === '') {
-                setSearchQuery(null)
-              }
-              onBlur?.(e)
-            }}
-          />
-        </Box>
+    const handleClearSearch = () => {
+      updateSearchQuery(null)
+      setSelectedValue(null)
+      onFound('')
+    }
 
-        {showOptions && (
-          <SearchOptions
-            displayedList={filteredList}
-            searchTerm={searchQuery || ''}
-            onSelect={handleSelect}
-            positionRelative={resultsRelativePosition}
-            resultsBorder={resultsBorder}
-            onNotFound={onNotFound}
-            notFoundComponent={notFoundComponent}
-          />
-        )}
-      </Field>
+    const handleCaretClick = () => {
+      setShowOptions(!showOptions)
+    }
+
+    const handleKeyDown = (event: {
+      key: string
+      preventDefault: () => void
+    }) => {
+      if (event.key === 'Enter' && highlightedIndex !== -1) {
+        event.preventDefault()
+        const focusedItem = filteredList[highlightedIndex]
+        handleSelect(focusedItem)
+      } else if (event.key === 'ArrowDown') {
+        event.preventDefault()
+        const nextIndex = (highlightedIndex + 1) % filteredList.length
+        setHighlightedIndex(nextIndex)
+      } else if (event.key === 'ArrowUp') {
+        event.preventDefault()
+        const prevIndex =
+          (highlightedIndex - 1 + filteredList.length) % filteredList.length
+        setHighlightedIndex(prevIndex)
+      }
+    }
+
+    const showClearSearchButton =
+      !!clearSearch && (!!value || !!selectedValue || !!searchQuery)
+
+    return (
+      <Wrapper ref={wrapperRef}>
+        <Field
+          className={className}
+          renderAsTitle={renderAsTitle}
+          htmlFor={id}
+          {...otherProps}
+        >
+          <Box flex alignItems="center" justifyContent="flex-start">
+            {showIcon && <StyledFrontIcon render="search" color="sesame" />}
+            <Input
+              id={id}
+              name={name}
+              ref={ref}
+              placeholder={placeholder}
+              $error={otherProps.error}
+              $frontIcon={showIcon ? 'search' : undefined}
+              $fallbackStyle={fallbackStyle}
+              autoComplete="off"
+              value={displayedInputText}
+              onFocus={handleClick}
+              onChange={handleInputChange}
+              selected={isSelected}
+              onClick={handleClick}
+              onKeyDown={handleKeyDown}
+              onBlur={(e) => {
+                onBlur?.(e)
+              }}
+            />
+            <Icons
+              flex
+              alignItems="center"
+              gap="8px"
+              $clearSearch={showClearSearchButton}
+            >
+              {showClearSearchButton && (
+                <IconStrict
+                  type="button"
+                  render="plus"
+                  rotate={45}
+                  iconColor="marzipan"
+                  handleClick={handleClearSearch}
+                  size={24}
+                />
+              )}
+              <Line />
+              <IconStrict
+                type="button"
+                render="caret"
+                iconColor="marzipan"
+                rotate={showOptions ? 180 : 0}
+                handleClick={handleCaretClick}
+                size={24}
+              />
+            </Icons>
+          </Box>
+
+          {showOptions && (
+            <SearchOptions
+              displayedList={filteredList}
+              selectedValue={selectedValue}
+              highlightedIndex={highlightedIndex}
+              setHighlightedIndex={setHighlightedIndex}
+              onKeyDown={handleKeyDown}
+              searchTerm={searchQuery || ''}
+              onSelect={handleSelect}
+              positionRelative={resultsRelativePosition}
+              resultsBorder={resultsBorder}
+              onNotFound={onNotFound}
+              notFoundComponent={
+                notFoundComponent && notFoundComponent(searchQuery ?? '')
+              }
+            />
+          )}
+        </Field>
+      </Wrapper>
     )
   },
 )
+
+const Wrapper = styled(Box)`
+  position: relative;
+`
+
+const Line = styled(Box)`
+  background: ${theme.colors.oatmeal};
+  height: 24px;
+  width: 1px;
+`
+
+export const Icons = styled(Box)<{ $clearSearch: boolean }>`
+  position: relative;
+  right: ${({ $clearSearch }) => ($clearSearch ? '80px' : '48px')};
+  margin-right: ${({ $clearSearch }) => ($clearSearch ? '-80px' : '-48px')};
+`